メンタルゲージ用UIを各キャラのメンタルゲージと連動＆現在量に応じて色が変わるようにした
