--- Assets/CameraManager.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: a9c68b7fb6320174c9077af8c6e0ea0f
-timeCreated: 1538637169
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
