--- Assets/CircleFill.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 58e8185339780344db489978feb29c23
-timeCreated: 1539230378
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
