--- Assets/Script/Tamura/Shout.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: c45b7565f8394da469a2fc9c257db472
-timeCreated: 1536885554
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
