--- Assets/Script/Tamura/voiceUpdate.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: e12486991984d3e47a1cac745d63e02a
-timeCreated: 1536887944
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
